requirements.txt
# Sort by a-z

cmake==3.18.4.post1
numpy==1.19.4
dlib==19.21.0
face_recognition==1.3.0
opencv-python==4.4.0.46

django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.3.0
django-model-controller==0.4.0
djangorestframework==3.11.1
ipdb==0.13.4
markdown==3.2.2
openpyxl==3.0.5
pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1